+++ src/test/scala/org/ergoplatform/nodeView/wallet/WalletScanLogicSpec.scala
@@ -35,7 +35,7 @@ class WalletScanLogicSpec extends ErgoPropertyTest with DBSpec with WalletTestOp
   private val trueProp = org.ergoplatform.settings.Constants.TrueLeaf
   private val scanningPredicate = EqualsScanningPredicate(ErgoBox.ScriptRegId, ByteArrayConstant(trueProp.bytes))
   private val appReq = ScanRequest("True detector", scanningPredicate, Some(ScanWalletInteraction.Off), None)
-  private val scanId: ScanId = ScanId @@ 50.toShort
+  private val scanId: ScanId = ScanId @@ 50
 
   private val pubkeys = prover.hdPubKeys
   private val miningScripts = WalletCache.miningScripts(pubkeys, s)
